\name{get.discrete.solution}
\alias{get.discrete.solution}
\title{l2boost internal method to get discrete solution (critical value, critical point)}
\usage{
  get.discrete.solution(rho.m, corr.x, lr, nu)
}
\description{
  l2boost internal method to get discrete solution
  (critical value, critical point)
}

