\name{get.hybrid.solution}
\alias{get.hybrid.solution}
\title{l2boost internal method to get hybrid solution (step size, critical value)}
\usage{
  get.hybrid.solution(rho.m, corr.x, lr)
}
\arguments{
  \item{rho.m}{vector of gradient corellations to this
  point (m)}

  \item{corr.x}{correlation matrix}

  \item{lr}{current step direction}
}
\description{
  l2boost internal method to get hybrid solution (step
  size, critical value)
}

