\docType{package}
\name{l2boost-package}
\alias{l2boost-package}
\title{l2boost - Boosting Coordinate Directions with L2-Loss}
\description{
  An boosting implementation using coordinate directions
  (design matrix columns) as basis functions and an l2-loss
  function. The l2boost package includes the elasticBoost
  extensions.
}
\references{
  John Ehrlinger, Hemant Ishwaran (2012). Characterizing
  l2boosting. Annals of Statists, to appear.
  \url{http://www.jstatsoft.org/v40/i03/}.
}

