\name{l2boost.default}
\alias{l2boost.default}
\title{default l2boost method takes a design matrix and response vector}
\usage{
  l2boost.default(x, y, M = NULL, nu = 1e-04,
    lambda = NULL, trace = FALSE,
    type = c("friedman", "discrete", "hybrid", "lars"),
    qr.tolerance = 1e-30,
    eps.tolerance = .Machine$double.eps, ...)
}
\description{
  default l2boost method takes a design matrix and response
  vector
}

