\name{l2boost.formula}
\alias{l2boost.formula}
\title{l2boost taking a formula object instead of design matrix.}
\usage{
  l2boost.formula(formula, data = list(), ...)
}
\arguments{
  \item{formula}{an object of class ‘"formula"’ (or one
  that can be coerced to that class): a symbolic
  description of the model to be fitted.  The details of
  model specification are given under "Details".}

  \item{data}{an optional data frame, list or environment
  (or object coercible by ‘as.data.frame’ to a data frame)
  containing the variables in the model.  If not found in
  ‘data’, the variables are taken from
  "environment(formula)", typically the environment from
  which "lm" is called.}
}
\description{
  l2boost taking a formula object instead of design matrix.
}

