\name{mstep.long}
\alias{mstep.long}
\title{l2boost function to determine step size for long descent
|Dk| must be less than or equal to 1; force it otherwise
enforce step=infty when dk=Rk=1}
\usage{
  mstep.long(rho.m, Rk, lr, nu)
}
\description{
  l2boost function to determine step size for long descent
  |Dk| must be less than or equal to 1; force it otherwise
  enforce step=infty when dk=Rk=1
}

