\name{nu.limit}
\alias{nu.limit}
\title{limiting nu size}
\usage{
  nu.limit(rho.m, Rk, lr)
}
\arguments{
  \item{rho.m}{vector of stagewise regression parameters}

  \item{Rk}{correlation coefficient of current and
  candidate coordinate directions}

  \item{lr}{candidate direction index}
}
\description{
  limiting nu size
}

