\name{plot.l2boost}
\alias{plot.l2boost}
\title{Plots l2boost objects}
\usage{
  plot.l2boost(obj, type = c("rho", "coef"),
    standardize = TRUE, active.set = NULL,
    xvar = c("step", "norm"), x.lab = NULL, y.lab = NULL,
    trim = TRUE, clip = NULL, col = NULL, ylim = NULL,
    xlim = NULL)
}
\value{
  \code{NULL}
}
\description{
  Plots l2boost objects
}

