\name{residuals.l2boost}
\alias{residuals.l2boost}
\title{residuals.l2boost}
\usage{
  residuals.l2boost(object, m = NULL, ...)
}
\arguments{
  \item{object}{an l2boost object for the extraction of
  model coefficients.'}

  \item{m}{the iteration number with the l2boost path. If
  m=NULL, the coefficients are obtained from the last
  iteration M.}

  \item{...}{arguments (unused)}
}
\value{
  \code{NULL}
}
\description{
  residuals.l2boost
}

