\name{summary.l2boost}
\alias{summary.l2boost}
\title{Unimplemented generic function
These are placeholders right now.}
\usage{
  summary.l2boost(object, ...)
}
\arguments{
  \item{object}{an l2boost object}

  \item{...}{other arguments (unused)}
}
\description{
  Unimplemented generic function These are placeholders
  right now.
}

