% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR}
\alias{VAR}
\title{This is a hidden function of the l2boost package.
VAR is a helper function that specifically returns NA if all 
values of the argument x are NA, otherwise, it returns a var 
object.}
\usage{
VAR(x)
}
\arguments{
\item{x}{return variance of x matrix.}
}
\description{
This is a hidden function of the l2boost package.
VAR is a helper function that specifically returns NA if all 
values of the argument x are NA, otherwise, it returns a var 
object.
}
