\name{which.max.ind}
\alias{which.max.ind}
\title{determines which index maximizes a vector, with some elements excluded}
\usage{
  which.max.ind(vec, exclude = NULL)
}
\arguments{
  \item{vec}{values to maximize}

  \item{exclude}{vector of indices to exclud}
}
\description{
  determines which index maximizes a vector, with some
  elements excluded
}

