\name{which.min.ind}
\alias{which.min.ind}
\title{determines which index minimizes a vector, with some elements excluded}
\usage{
  which.min.ind(vec, exclude = NULL)
}
\arguments{
  \item{vec}{vector of values to minimize}

  \item{exclude}{vector of indices to exclude}
}
\description{
  determines which index minimizes a vector, with some
  elements excluded
}

