\name{coef.l2boost}
\alias{coef.l2boost}
\title{coef.l2boost}
\usage{
  coef.l2boost(object, m = NULL, ...)
}
\arguments{
  \item{object}{an l2boost object for the extraction of
  model coefficients.'}

  \item{m}{the iteration number with the l2boost path. If
  m=NULL, the coefficients are obtained from the last
  iteration M.}

  \item{...}{generic arguments passed to parent function}
}
\description{
  coef.l2boost
}

