\name{get.discrete.solution}
\alias{get.discrete.solution}
\title{l2boost internal method to get discrete step sized solution (critical value, critical point)}
\usage{
  get.discrete.solution(rho.m, corr.x, lr, nu)
}
\arguments{
  \item{rho.m}{vector of gradient corellations to this
  point (m)}

  \item{corr.x}{correlation matrix}

  \item{lr}{current step direction}

  \item{nu}{l1 shrinkage parameter}
}
\description{
  l2boost internal method to get discrete step sized
  solution (critical value, critical point)
}

