% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.l2boost.R
\name{print.summary.l2boost}
\alias{print.summary.l2boost}
\title{Unimplemented generic function
These are placeholders right now.}
\usage{
\method{print}{summary.l2boost}(x, ...)
}
\arguments{
\item{x}{an l2boost object}

\item{...}{other arguments (not used)}
}
\description{
Unimplemented generic function
These are placeholders right now.
}
