% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.l2boost.R
\name{summary.l2boost}
\alias{summary.l2boost}
\title{Unimplemented generic function
These are placeholders right now.}
\usage{
\method{summary}{l2boost}(object, ...)
}
\arguments{
\item{object}{an l2boost object}

\item{...}{other arguments (unused)}
}
\description{
Unimplemented generic function
These are placeholders right now.
}
